HW1_DynamicConv/requirement.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.1
grpcio==1.71.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.0.3
Pillow==10.0.0
protobuf==6.30.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.12.2
six==1.17.0
sympy==1.13.3
tensorboard==1.12.0
threadpoolctl==3.6.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.65.0
triton==2.2.0
typing_extensions==4.13.0
tzdata==2025.2
Werkzeug==3.1.3
